add functionality to receive current geolocation coordinates
On the page, we are still displaying the starting markup. In the console the user will be able to see a google map link with the current coordinates populated in it.
Clicking on this link will open up Google Maps in an adjacent browser tab to display the user's current location.